requirements.txt
import-ipynb==0.1.4
ipykernel==6.19.2
ipython==8.7.0
joblib==1.2.0
mysql-connector-python==8.0.32
numpy==1.23.5
openpyxl==3.0.10
pandas==1.3.1
selenium==4.7.2
texthero==1.1.0
nltk==3.7
scikit-learn==1.0.2
webdriver-manager==3.8.5
wordcloud==1.8.2.2
PyAutoGUI==0.9.53
psutil==5.9.4
schedule==1.2.0
apache-airflow==2.6.1
spacy == 2.3.9
emoji == 2.2.0
selenium==4.7.2
gensim==3.8.1
fuzzywuzzy==0.18.0
jupyter==1.0.0
kmodes==0.12.1
matplotlib==3.5.0
plotly==5.2.1
seaborn==0.11.2
xgboost==1.5.0
boto3==1.18.8
python-decouple==3.4